@vpaprotsk 2025-04-03
> > > I think it would also be useful to write a release note describing the approximate performance improvement gains for the crypto algorithms as displayed in your chart. Thanks.
> > 
> > 
> > @seanjmullan I think I only done that once, cant find the 'instructions'.. I think Jamil had helped me, but.. (https://bugs.openjdk.org/browse/JDK-8297970) "Create subtask with 'release-note' label?"
> 
> See the [Release Notes section](https://openjdk.org/guide/#release-notes) of the OpenJDK Developer's Guide for the process.
> 
> For this release note, something similar to https://jdk.java.net/24/release-notes#JDK-8333867 would be nice, a couple of sentences explaining the approximate improvement, on what architectures, and for what APIs and algorithms one would see the improvement.

Done I think: https://bugs.openjdk.org/browse/JDK-8297970