@seanjmullan 2025-04-04
> Done I think: https://bugs.openjdk.org/browse/JDK-8297970

Is this link correct? This issue was fixed in JDK 20.