@seanjmullan 2025-04-08
> > > Done I think: https://bugs.openjdk.org/browse/JDK-8297970
> > 
> > 
> > Is this link correct? This issue was fixed in JDK 20.
> 
> Sorry.. copy/paste didnt notice.. https://bugs.openjdk.org/browse/JDK-8353670 (also ends in *70!)

Looks good, but can you also say something about the approximate performance gain?